:recycle: Move injector
